requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
DingtalkChatbot==1.5.2
Django==3.1.5
django-bootstrap4==2.3.1
django-celery==3.3.1
django-filter==2.4.0
django-grappelli==2.14.3
django-mdeditor==0.1.18
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-python3-ldap==0.11.4
django-redis==4.12.1
django-registration-redux==2.9
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==2.1.1
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==3.0.37
ldap3==2.9
Markdown==3.3.3
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.20.1
opencv-python==4.5.1.48
pika==1.2.0
Pillow==8.1.0
prompt-toolkit==3.0.16
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
